📝 chore(.gitignore): add .gitignore file to ignore unnecessary files and directories
This commit adds a .gitignore file to the project. The .gitignore file is used to specify files and directories that should be ignored by Git. In this commit, the following files and directories are added to the .gitignore file:

- Nuxt dev/build outputs: .output, .data, .nuxt, .nitro, .cache, dist
- Node dependencies: node_modules
- Logs: logs, *.log
- Misc: .DS_Store, .fleet, .idea
- Local env files: .env, .env.* (except .env.example)

By ignoring these files and directories, we can keep our repository clean and avoid committing unnecessary files and sensitive information.