📦 feat(TodoCardCreate.vue): add TodoCardCreate component to allow users to create new todos with an image and save them to local storage
This commit adds the `TodoCardCreate.vue` component, which is responsible for rendering a form to create new todos. The component includes a `TodoImage` component to allow users to select an image for the todo. The selected image ID is stored in the `imageId` variable using the `setImageId` function.

When the form is submitted, the `saveTodoOnLocalStorage` function is called. This function generates a unique ID for the todo, combines the form data with the selected image ID, and saves the todo to local storage using the `saveItemOnLocalStorage` function.

The form includes fields for the todo title and description, with appropriate validation. The form submission button is labeled "Criar Todo" and has a green background color.

Note: There is no newline at the end of the file.